Add android module to project.
This adds in the android module to the to-do list case study project.
This project will follow the Clean Architecture design, and will use the
following to assist in the creation.
  - Dagger 2 for dependency injection
  - Butterknife for view injection
  - RxJava & Retrolambda to be reactive
  - Retrofit2 for a wonderfully light, and intuitive networking library

This will showcase how a client must interact with a REST-ful service
from start to finish, and I hope to showcase some client specific struggles that will help backend engineers understand the paradigms that client engineers face.