Set up scala modules.
This is the basic setup I will (hopefully) be using for my
microarchitecture. I wanted to see how clean architecture works outside
of java since theoretically it should work in any project. I believe
this will actually help us with dealing with any couplings that could
arise especially if I am required further on to add in any java to this
project.